requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
beautifulsoup4==4.11.1
black==23.12.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.7
Django==4.0.6
django-cors-headers==4.3.1
django-countries==7.4.2
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
html5lib==1.1
idna==3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.0
platformdirs==4.1.0
promise==2.3
python-dotenv==1.0.1
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.12
webencodings==0.5.1
